requirements.txt
flask
pika
pymongo
python-dateutil